Cleanup and version bump
