+++ domains/data-sync/chart/Chart.yaml
@@ -3,4 +3,4 @@ name: eo-data-sync
 description: A chart containing the sync domain
 
 type: application
-version: 0.4.0
+version: 0.5.0
