+++ domains/emissions/chart/Chart.yaml
@@ -3,7 +3,7 @@ name: eo-emissions
 description: A chart containing the emissions domain.
 
 type: application
-version: 0.0.5
+version: 0.1.0
 
 dependencies:
   - name: eo-base-helm-chart
