+++ domains/emissions/chart/values.yaml
@@ -10,7 +10,7 @@ eo-base-helm-chart:
       probes: true
       image:
         repository: ghcr.io/energinet-datahub/eo-emissions
-        tag: 0.0.5
+        tag: 0.1.0
   service:
     deployment: api
     type: ClusterIP
